seperating gulp tasks in the gulp directory very helpfull for bigger projects
