+++ gulp/pagespeed.js
@@ -1,0 +1,16 @@
+'use strict';
+
+// Include Gulp & Tools We'll Use
+var gulp = require('gulp');
+var pagespeed = require('psi');
+
+// Run PageSpeed Insights
+// Update `url` below to the public URL for your site
+gulp.task('pagespeed', pagespeed.bind(null, {
+  // By default, we use the PageSpeed Insights
+  // free (no API key) tier. You can use a Google
+  // Developer API key if you have one. See
+  // http://goo.gl/RkN0vE for info key: 'YOUR_API_KEY'
+  url: 'https://example.com',
+  strategy: 'mobile'
+}));
