+++ gulp/serve_dist.js
@@ -1,0 +1,18 @@
+'use strict';
+
+// Include Gulp & Tools We'll Use
+var gulp = require('gulp');
+var browserSync = require('browser-sync');
+
+// Build and serve the output from the dist build
+gulp.task('serve:dist', ['default'], function () {
+  browserSync({
+    notify: false,
+    logPrefix: 'WSK',
+    // Run as an https by uncommenting 'https: true'
+    // Note: this uses an unsigned certificate which on first access
+    //       will present a certificate warning in the browser.
+    // https: true,
+    server: 'dist'
+  });
+});
